+++ common/src/main/java/org/infernalstudios/questlog/core/QuestManager.java
@@ -167,6 +167,10 @@ public class QuestManager {
                     quest.hasSentCompletion = true;
                     QuestlogEvents.onQuestCompleted(new QuestEvent.Completed(this.player, quest, true));
                     Questlog.LOGGER.trace("Sent quest completed event for {}", id);
+
+                    if (quest.isGlobal() && ServerPlayerManager.INSTANCE != null) {
+                        ServerPlayerManager.INSTANCE.onGlobalQuestUpdated(quest);
+                    }
                 }
             }
         }