+++ common/src/main/java/org/infernalstudios/questlog/core/quests/objectives/logic/NotObjective.java
@@ -30,9 +30,9 @@ public class NotObjective extends Objective {
     }
 
     @Override
-    public void markAsRequirement() {
-        super.markAsRequirement();
-        this.child.markAsRequirement();
+    public void markAsPrerequisite() {
+        super.markAsPrerequisite();
+        this.child.markAsPrerequisite();
     }
 
     @Override