+++ common/src/main/java/org/infernalstudios/questlog/core/quests/rewards/CommandReward.java
@@ -25,7 +25,13 @@ public class CommandReward extends Reward {
                 .withPermission(this.permissionLevel)
                 .withSuppressedOutput();
 
-        player.getServer().getCommands().performPrefixedCommand(source, this.command);
+        String formattedCommand = this.command
+                .replace("{player}", player.getGameProfile().getName())
+                .replace("%player%", player.getGameProfile().getName())
+                .replaceAll("(?<!\\w)@p(?!\\w)", player.getGameProfile().getName())
+                .replaceAll("(?<!\\w)@s(?!\\w)", player.getGameProfile().getName());
+
+        player.getServer().getCommands().performPrefixedCommand(source, formattedCommand);
 
         super.applyReward(player);
     }